+++ tests/codegen.rs
@@ -38,7 +38,7 @@ fn task_with_options() -> String {
 
 #[test]
 fn test_task_options() {
-    let t = task_with_options::new();
+    let t = task_with_options {};
     assert_eq!(t.timeout(), Some(2));
     assert_eq!(t.max_retries(), Some(3));
     assert_eq!(t.min_retry_delay(), Some(0));
@@ -65,8 +65,8 @@ fn task_with_strings(s1: String, s2: String) -> String {
 
 #[test]
 fn test_task_with_strings() {
-    let s = task_with_strings::s("hi".into(), "there".into());
-    let t = task_with_strings::new();
-    let result = futures::executor::block_on(t.run(s)).unwrap();
+    let s = task_with_strings::new("hi".into(), "there".into());
+    let t = task_with_strings {};
+    let result = futures::executor::block_on(t.run(s.params)).unwrap();
     assert_eq!(result, "hi, there");
 }